Multiple Changes to Combo logic:
Reworked Combo logic to work for BrineBreaker without the need for combo logic to be repeated within its class.
Reworked Combo Events to be triggered when entities are damaged (or die), instead of when an entity is hit.
Revamp of the Combo UI, stopped using texture and is now only using text, and UI now appears and disappears with a slower animation.